to prod assistant bot

+++ src/components/AssistantBot/index.tsx
@@ -6,6 +6,7 @@ import {
   DialogContent,
   DialogTrigger,
 } from "@site/src/components/ui/dialog";
+import EmailInputDialog from "../ui/EmailInputDialog";
 import ChatGPTLogo from "@site/static/img/chatgpt.webp";
 import useAssistantBot from "@site/src/hooks/useAssistantBot";
 import { Avatar, AvatarImage } from "@site/src/components/ui/avatar";
@@ -28,16 +29,53 @@ const chatHistoryKey = "chatHistory";
 
 export default function AssistantBot() {
   const [query, setQuery] = useState("");
-  const { answer, generateAnswer, loading, reset } = useAssistantBot();
+  const { answer, generateAnswer, loading, reset, error } = useAssistantBot();
   const [messages, setMessages] = useState<Message[]>([]);
   const [threadId, setThreadId] = useState(null);
   const [status, setStatus] = useState("");
+  const [botError, SetBotError] = useState("");
 
   const [open, setOpen] = useState(false);
   const messagesEndRef = useRef(null);
   const scrollToBottom = () => {
     messagesEndRef.current.scrollIntoView({ behavior: "smooth" });
   };
+
+  useEffect(() => {
+    if (error) {
+      SetBotError(error);
+    }
+  }, [error]);
+
+  // Function to set data with an expiration time
+  function setWithExpiry(key, value, ttl) {
+    const now = new Date();
+    // 'item' is an object which contains the original value as well as the expiration time
+    const item = {
+      value: value,
+      expiry: now.getTime() + ttl,
+    };
+    localStorage.setItem(key, JSON.stringify(item));
+  }
+
+  // Function to get data with checking the expiration time
+  function getWithExpiry(key) {
+    const itemStr = localStorage.getItem(key);
+    // If the item doesn't exist, return null
+    if (!itemStr) {
+      return null;
+    }
+    const item = JSON.parse(itemStr);
+    const now = new Date();
+    // Compare the expiry time of the item with the current time
+    if (now.getTime() > item.expiry) {
+      // If the item is expired, delete the item from storage and return null
+      localStorage.removeItem(key);
+      return null;
+    }
+    return item.value;
+  }
+
   useEffect(() => {
     if (messages.length > 0 && messagesEndRef.current) scrollToBottom();
   }, [messages]);
@@ -57,6 +95,7 @@ export default function AssistantBot() {
       }
       const data = await response.json();
       localStorage.setItem("thread_id", data.id); // Save thread ID to local storage
+      setWithExpiry("chatExpired", "valid", 24 * 60 * 60 * 1000); // Set the expiry to 24 hours
       setStatus("");
       setThreadId(data.id); // Save thread ID to state
     } catch (error) {
@@ -64,30 +103,39 @@ export default function AssistantBot() {
     }
   }
   async function deleteThreadId() {
+    const storedThreadId = localStorage.getItem("thread_id");
     try {
       setStatus("Resetting...");
-      const response = await fetch(
-        `/api/moralis-assistant-thread-delete?thread_id=${threadId}`,
-        {
-          method: "DELETE",
+      if (storedThreadId) {
+        const response = await fetch(
+          `/api/moralis-assistant-thread-delete?thread_id=${storedThreadId}`,
+          {
+            method: "DELETE",
+          }
+        );
+        if (!response.ok) {
+          setStatus("Failed to reset. Please reload the page.");
+          throw new Error("Network response was not ok");
         }
-      );
-      if (!response.ok) {
-        setStatus("Failed to reset. Please reload the page.");
-        throw new Error("Network response was not ok");
       }
       await fetchThreadId();
     } catch (error) {
       console.error("Failed to fetch thread ID:", error);
     }
   }
+
   useEffect(() => {
-    // Check if the thread_id is already saved in local storage
-    const storedThreadId = localStorage.getItem("thread_id");
-    if (!storedThreadId) {
-      fetchThreadId(); // Fetch and save thread ID if not already saved
+    const chatExpired = getWithExpiry("chatExpired");
+    if (chatExpired) {
+      // Check if the thread_id is already saved in local storage
+      const storedThreadId = localStorage.getItem("thread_id");
+      if (!storedThreadId) {
+        fetchThreadId(); // Fetch and save thread ID if not already saved
+      } else {
+        setThreadId(storedThreadId); // Use the saved thread ID
+      }
     } else {
-      setThreadId(storedThreadId); // Use the saved thread ID
+      deleteThreadId();
     }
   }, []);
 
@@ -196,145 +244,200 @@ export default function AssistantBot() {
     setQuery("");
   };
 
+  const [isEmailDialogOpen, setEmailDialogOpen] = useState(false);
+
+  const openEmailDialog = () => {
+    if (!localStorage.getItem("userEmail")) {
+      setEmailDialogOpen(true);
+    }
+  };
+
+  useEffect(() => {
+    openEmailDialog();
+  }, []);
+
+  const closeEmailDialog = () => {
+    setEmailDialogOpen(false);
+    if (!localStorage.getItem("userEmail")) {
+      setOpen(false);
+    }
+  };
+
+  useEffect(() => {
+    if (open) {
+      openEmailDialog();
+    }
+  }, [open]);
+
   return (
-    <Dialog open={open} onOpenChange={setOpen}>
-      <DialogTrigger asChild>
-        <Avatar>
-          <AvatarImage src={ChatGPTLogo} alt="ChatGPT" />
-        </Avatar>
-      </DialogTrigger>
-      <DialogContent
-        aria-label="Chat with AI"
-        className="fixed inset-x-0 top-10 bottom-10 mx-auto flex justify-center items-center p-4"
-        style={{ minWidth: "60vw" }} // Set the inline style for large screens
-      >
-        <div className="bg-white dark:bg-slate-800 rounded-lg shadow-lg flex flex-col w-full max-w-5xl h-full my-8">
-          <div className="p-4 flex items-center justify-between">
-            <h2 className="text-2xl font-semibold text-white">Chat with AI</h2>
-            <button
-              onClick={handleReset}
-              className="text-white inline-flex items-center"
-            >
-              <RotateCcw size={20} className="mr-2 text-white" />
-              Reset
-            </button>
-          </div>
-          <ScrollArea className="flex-grow overflow-y-auto">
-            <div className="flex flex-col p-4 space-y-2">
-              {/* Reduced space between messages */}
-              {messages.map((message, index) => (
-                <div
-                  key={index}
-                  className={`flex ${
-                    message.role === "assistant"
-                      ? "justify-start"
-                      : "justify-end"
-                  }`}
-                >
-                  {message.role === "assistant" ? (
-                    <div className="flex items-start w-5/6">
-                      <Avatar className="flex-shrink-0">
-                        <AvatarImage
-                          src={ChatGPTLogo}
-                          alt="ChatGPT"
-                          className="w-8 h-8"
-                        />
-                      </Avatar>
-                      <div className="ml-4 bg-slate-300 dark:bg-slate-600 p-2 rounded-lg">
-                        <p className="text-sm dark:text-white mb-0">
+    <>
+      <Dialog open={open} onOpenChange={setOpen}>
+        <DialogTrigger asChild>
+          <Avatar>
+            <AvatarImage src={ChatGPTLogo} alt="ChatGPT" />
+          </Avatar>
+        </DialogTrigger>
+        <DialogContent
+          aria-label="Chat with AI"
+          className="fixed inset-x-0 top-10 bottom-10 mx-auto flex justify-center items-center p-4"
+          style={{ minWidth: "60vw" }} // Set the inline style for large screens
+        >
+          <div className="bg-white dark:bg-slate-800 rounded-lg shadow-lg flex flex-col w-full max-w-5xl h-full my-8">
+            <div className="p-4 flex items-center justify-between">
+              <h2 className="text-2xl font-semibold text-white">
+                Chat with AI
+              </h2>
+              <button
+                onClick={handleReset}
+                className="text-white inline-flex items-center"
+              >
+                <RotateCcw size={20} className="mr-2 text-white" />
+                Reset
+              </button>
+            </div>
+            <ScrollArea className="flex-grow overflow-y-auto">
+              <div className="flex flex-col p-4 space-y-2">
+                {/* Reduced space between messages */}
+                {messages.map((message, index) => (
+                  <div
+                    key={index}
+                    className={`flex ${
+                      message.role === "assistant"
+                        ? "justify-start"
+                        : "justify-end"
+                    }`}
+                  >
+                    {message.role === "assistant" ? (
+                      <div className="flex items-start w-5/6">
+                        <Avatar className="flex-shrink-0">
+                          <AvatarImage
+                            src={ChatGPTLogo}
+                            alt="ChatGPT"
+                            className="w-8 h-8"
+                          />
+                        </Avatar>
+                        <div className="ml-4 bg-slate-300 dark:bg-slate-600 p-2 rounded-lg">
+                          <p className="text-sm dark:text-white mb-0">
+                            <ReactMarkdown remarkPlugins={[remarkGfm]}>
+                              {cleanMessage(message.content)}
+                            </ReactMarkdown>
+                          </p>
+                        </div>
+                      </div>
+                    ) : (
+                      <div className="bg-blue-500 p-2 rounded-lg max-w-4/6">
+                        <p className="text-white mb-0">
                           <ReactMarkdown remarkPlugins={[remarkGfm]}>
                             {cleanMessage(message.content)}
                           </ReactMarkdown>
                         </p>
                       </div>
-                    </div>
-                  ) : (
-                    <div className="bg-blue-500 p-2 rounded-lg max-w-4/6">
-                      <p className="text-white mb-0">
+                    )}
+                  </div>
+                ))}
+                {loading && (
+                  <div className="flex items-start w-5/6">
+                    <Avatar className="flex-shrink-0">
+                      <AvatarImage
+                        src={ChatGPTLogo}
+                        alt="ChatGPT"
+                        className="w-8 h-8"
+                      />
+                    </Avatar>
+                    <div className="ml-4 bg-slate-300 dark:bg-slate-600 p-2 rounded-lg">
+                      <p className="text-sm dark:text-white mb-0">
                         <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                          {cleanMessage(message.content)}
+                          Loading...
                         </ReactMarkdown>
                       </p>
                     </div>
-                  )}
-                </div>
-              ))}
-              {loading && (
-                <div className="flex items-start w-5/6">
-                  <Avatar className="flex-shrink-0">
-                    <AvatarImage
-                      src={ChatGPTLogo}
-                      alt="ChatGPT"
-                      className="w-8 h-8"
-                    />
-                  </Avatar>
-                  <div className="ml-4 bg-slate-300 dark:bg-slate-600 p-2 rounded-lg">
-                    <p className="text-sm dark:text-white mb-0">
-                      <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                        Loading...
-                      </ReactMarkdown>
-                    </p>
+                  </div>
+                )}
+                {status && (
+                  <div className="flex justify-center">
+                    <div>{status}</div>
+                  </div>
+                )}
+                {error && (
+                  <div className="flex items-start w-5/6">
+                    <Avatar className="flex-shrink-0">
+                      <AvatarImage
+                        src={ChatGPTLogo}
+                        alt="ChatGPT"
+                        className="w-8 h-8"
+                      />
+                    </Avatar>
+                    <div className="ml-4 bg-slate-300 dark:bg-slate-600 p-2 rounded-lg">
+                      <p className="text-sm dark:text-orange-400 mb-0">
+                        <ReactMarkdown remarkPlugins={[remarkGfm]}>
+                          Error occured in answering. Please refresh the page or
+                          contact customer support.
+                        </ReactMarkdown>
+                      </p>
+                    </div>
+                  </div>
+                )}
+                {messages && (
+                  <div ref={messagesEndRef} onLoad={scrollToBottom} />
+                )}
+              </div>
+            </ScrollArea>
+            <div className="p-4 flex-none">
+              <form onSubmit={handleUserSubmit} className="flex items-center">
+                <input
+                  type="text"
+                  className="w-full p-3 text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white"
+                  placeholder="Type your question here..."
+                  value={query}
+                  onChange={handleTextChange}
+                  required
+                  style={{ minHeight: "50px" }}
+                />
+                <button
+                  type="submit"
+                  className="ml-4 bg-blue-500 text-white py-2 px-4 rounded text-lg"
+                >
+                  Send
+                </button>
+              </form>
+            </div>
+            <div className="p-4 flex-none">
+              <Alert className="flex items-center justify-between">
+                <div className="flex items-center">
+                  <Microscope className="h-6 w-6 text-yellow-500" />
+                  &emsp;
+                  <div>
+                    <AlertTitle className="text-sm font-semibold">
+                      Moralis AI is experimental can give wrong or outdated
+                      data.
+                    </AlertTitle>
+                    <AlertDescription className="text-xs">
+                      Verify outputs or contact support if unsure.
+                    </AlertDescription>
                   </div>
                 </div>
-              )}
-              {status && (
-                <div className="flex justify-center">
-                  <div>{status}</div>
-                </div>
-              )}
-              {messages && <div ref={messagesEndRef} onLoad={scrollToBottom} />}
+                <button
+                  className="bg-gray-800 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded"
+                  onClick={() => {
+                    setOpen(false);
+                    document
+                      .getElementsByClassName("intercom-launcher")[0]
+                      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                      //@ts-ignore
+                      .click();
+                  }}
+                >
+                  Contact
+                </button>
+              </Alert>
             </div>
-          </ScrollArea>
-          <div className="p-4 flex-none">
-            <form onSubmit={handleUserSubmit} className="flex items-center">
-              <input
-                type="text"
-                className="w-full p-3 text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white"
-                placeholder="Type your question here..."
-                value={query}
-                onChange={handleTextChange}
-                required
-                style={{ minHeight: "50px" }}
-              />
-              <button
-                type="submit"
-                className="ml-4 bg-blue-500 text-white py-2 px-4 rounded text-lg"
-              >
-                Send
-              </button>
-            </form>
           </div>
-          <div className="p-4 flex-none">
-            <Alert className="flex items-center justify-between">
-              <div className="flex items-center">
-                <Microscope className="h-6 w-6 text-yellow-500" />
-                &emsp;
-                <div>
-                  <AlertTitle className="text-sm font-semibold">
-                    Moralis AI is experimental.
-                  </AlertTitle>
-                  <AlertDescription className="text-xs">
-                    Verify outputs or contact support if unsure.
-                  </AlertDescription>
-                </div>
-              </div>
-              <button
-                className="bg-gray-800 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded"
-                onClick={() => {
-                  setOpen(false);
-                  document
-                    .getElementsByClassName("intercom-launcher")[0]
-                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                    //@ts-ignore
-                    .click();
-                }}
-              >
-                Contact
-              </button>
-            </Alert>
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
+        </DialogContent>
+      </Dialog>
+      <EmailInputDialog
+        isOpen={open && isEmailDialogOpen}
+        onClose={closeEmailDialog}
+      />
+    </>
   );
 }
